updated unit test

+++ tests/unittest/export_gcs/test_gcs_hooks.py
@@ -12,20 +12,24 @@ class Node:
 
 class TestRunner(unittest.TestCase):
     def setUp(self):
-        os.environ["GCP_SERVICE_ACCOUNT"] = ""
-        self.runner = GCSExport()
-        self.original_item = dl.items.get(item_id='658ae4cd160fb30cdebf1156')
-        self.original_annotations = self.original_item.annotations.list()
+        item_id = "658ae4cd160fb30cdebf1156"
+        bucket_name = "micha-storage"
         remote_filepath = "/clones/1.jpg"
+
+        # Connect .env file
+        if os.environ.get("GCP_SERVICE_ACCOUNT") is None:
+            raise ValueError("Missing GCP service account json.")
+        self.runner = GCSExport()
+        self.original_item = dl.items.get(item_id=item_id)
+        self.original_annotations = self.original_item.annotations.list()
         try:
             item = self.original_item.dataset.items.get(filepath=remote_filepath)
             item.delete()
         except dl.exceptions.NotFound:
             pass
-
         self.item = self.original_item.clone(remote_filepath=remote_filepath)
         self.context = dl.Context()
-        self.context._node = Node(metadata={'customNodeConfig': {'bucket_name': 'micha-storage'}})
+        self.context._node = Node(metadata={'customNodeConfig': {'bucket_name': bucket_name}})
 
     def test_export_annotations(self):
         self.runner.export_annotation(item=self.item, context=self.context)
